Move encyclopedia data structure into common library
Cut down cross references between UI, parser and common code modules
