+++ universe/Encyclopedia.h
@@ -6,7 +6,9 @@
 #include <map>
 #include <vector>
 
-struct EncyclopediaArticle {
+#include "../util/Export.h"
+
+struct FO_COMMON_API EncyclopediaArticle {
     EncyclopediaArticle(const std::string& name_, const std::string& category_,
                         const std::string& short_description_, const std::string& description_,
                         const std::string& icon_) :
@@ -23,9 +25,11 @@ struct EncyclopediaArticle {
     std::string icon;
 };
 
-struct Encyclopedia {
+struct FO_COMMON_API Encyclopedia {
     Encyclopedia();
     std::map<std::string, std::vector<EncyclopediaArticle> >   articles;
 };
+
+FO_COMMON_API const Encyclopedia& GetEncyclopedia();
 
 #endif
